Create Node factory function
This was also a refresher for me with regards to scope. I initially
created the methods using the arrow function syntax, however, this
creates lexical scope. This means that when the method is called, the
scope is inherited from the context that the method was called. Instead,
I decided to use the regular function syntax for my methods, and use the
"this" keyword to make it so that the context of the properties being
changed is referring to the properties of the created object.